requirements.txt
vobject
delegator.py
click
cairosvg
pystache
loguru
